requirements.txt
Django==1.11.7
django-bower==5.2.0
djangorestframework==3.7.7
markdown==2.6.10
django-filter==1.1.0
python-bitcoinrpc==1.0
